contracts/lib/python/authmw/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "authmw"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["PyJWT>=2.8"]

[project.optional-dependencies]
dev = ["pytest>=7.4"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
